Simplified legend generating code. Using a single dictionary.
