--- scripts/scatterplots/plot-legends/legend_climate_community_resilience.json
@@ -1,172 +1,0 @@
-[
-    {
-        "idx": 1,
-        "key": "climate_community_resilience_1",
-        "subcat_en": "Community preparedness",
-        "color": "#5ed0b9",
-        "comment_ids": [
-            550,
-            262,
-            428,
-            624,
-            873,
-            890,
-            891,
-            906,
-            916,
-            1022,
-            1048,
-            1506,
-            1608,
-            287
-        ]
-    },
-    {
-        "idx": 2,
-        "key": "climate_community_resilience_2",
-        "subcat_en": "Fire resistant building and landscaping",
-        "color": "#70b6e5",
-        "comment_ids": [
-            997,
-            1218,
-            1572,
-            758,
-            274,
-            300,
-            301,
-            370,
-            379,
-            395,
-            471,
-            541,
-            640,
-            645,
-            658,
-            661,
-            673,
-            689,
-            693,
-            739,
-            751,
-            774,
-            783,
-            817,
-            918,
-            921,
-            928,
-            944,
-            962,
-            1021,
-            1050,
-            1147,
-            1160,
-            1174,
-            1239,
-            1247,
-            1260,
-            1266,
-            1273,
-            1302,
-            1327,
-            1342,
-            1369,
-            1383,
-            1406,
-            1418,
-            1422,
-            1429,
-            1474,
-            1479,
-            1490,
-            1545,
-            1612,
-            1623,
-            1635,
-            1649,
-            1666,
-            1678,
-            1704
-        ]
-    },
-    {
-        "idx": 3,
-        "key": "climate_community_resilience_3",
-        "subcat_en": "Fire-safe infrastructure management",
-        "color": "#b98acb",
-        "comment_ids": [
-            1293,
-            1698,
-            618,
-            719,
-            987,
-            1200,
-            512,
-            617,
-            641,
-            854,
-            1558,
-            583,
-            258,
-            626,
-            695,
-            703,
-            801,
-            1092,
-            1233,
-            1460,
-            865,
-            734,
-            1182,
-            362,
-            593,
-            722,
-            1301,
-            1527,
-            1537,
-            1556,
-            1586,
-            1722,
-            1734,
-            400
-        ]
-    },
-    {
-        "idx": 4,
-        "key": "climate_community_resilience_4",
-        "subcat_en": "Other",
-        "color": "lightgray",
-        "comment_ids": [
-            520,
-            1357,
-            1650,
-            1636,
-            515,
-            925,
-            1653,
-            1514,
-            1320,
-            1694,
-            259,
-            542,
-            437,
-            571,
-            625,
-            627,
-            672,
-            769,
-            910,
-            914,
-            917,
-            954,
-            975,
-            1227,
-            1317,
-            1440,
-            1445,
-            1464,
-            1496,
-            1603,
-            1712
-        ]
-    }
-]
