--- scripts/scatterplots/plot-legends/legend_economic_recovery.json
@@ -1,126 +1,0 @@
-[
-    {
-        "idx": 1,
-        "key": "economic_recovery_1",
-        "subcat_en": "Grant distribution oversight",
-        "color": "#5ed0b9",
-        "comment_ids": [
-            267,
-            276,
-            609,
-            653,
-            668,
-            704,
-            994,
-            1057,
-            1207,
-            1215,
-            1314,
-            1435,
-            1455,
-            1482,
-            1512,
-            1517,
-            1560,
-            1596,
-            1687
-        ]
-    },
-    {
-        "idx": 2,
-        "key": "economic_recovery_2",
-        "subcat_en": "Importance of supporting local businesses",
-        "color": "#70b6e5",
-        "comment_ids": [
-            299,
-            327,
-            354,
-            356,
-            369,
-            372,
-            414,
-            441,
-            505,
-            565,
-            684,
-            698,
-            702,
-            752,
-            760,
-            766,
-            829,
-            858,
-            875,
-            887,
-            937,
-            1049,
-            1077,
-            1078,
-            1146,
-            1161,
-            1177,
-            1185,
-            1194,
-            1210,
-            1300,
-            1340,
-            1392,
-            1495,
-            1542,
-            1569,
-            1613,
-            1641,
-            1672,
-            1710,
-            1714,
-            1735
-        ]
-    },
-    {
-        "idx": 3,
-        "key": "economic_recovery_3",
-        "subcat_en": "Incentives for small business recovery",
-        "color": "#b98acb",
-        "comment_ids": [
-            445,
-            540,
-            692,
-            745,
-            819,
-            974,
-            1051,
-            1063,
-            1334,
-            1364,
-            1395,
-            1614,
-            1630,
-            1658
-        ]
-    },
-    {
-        "idx": 4,
-        "key": "economic_recovery_4",
-        "subcat_en": "Other",
-        "color": "lightgray",
-        "comment_ids": [
-            288,
-            307,
-            465,
-            543,
-            552,
-            780,
-            861,
-            901,
-            966,
-            1091,
-            1287,
-            1384,
-            1414,
-            1450,
-            1471,
-            1544,
-            1588
-        ]
-    }
-]
