--- scripts/scatterplots/plot-legends/legend_housing_rebuilding.json
@@ -1,179 +1,0 @@
-[
-    {
-        "idx": 1,
-        "key": "housing_rebuilding_1",
-        "subcat_en": "ADU considerations",
-        "color": "#5ed0b9",
-        "comment_ids": [
-            368,
-            459,
-            506,
-            546,
-            598,
-            652,
-            770,
-            834,
-            952,
-            1154,
-            1168,
-            1201,
-            1212,
-            1230,
-            1375,
-            1619,
-            1674,
-            1736
-        ]
-    },
-    {
-        "idx": 2,
-        "key": "housing_rebuilding_2",
-        "subcat_en": "Affordability, density, and fire-safe rebuilding",
-        "color": "#70b6e5",
-        "comment_ids": [
-            264,
-            290,
-            479,
-            984,
-            1262,
-            1286,
-            355,
-            602,
-            714,
-            1093,
-            1315,
-            1563,
-            1659,
-            1706,
-            1720,
-            297,
-            404,
-            594,
-            622,
-            647,
-            654,
-            884,
-            893,
-            999,
-            1014,
-            1067,
-            1086,
-            1195,
-            1223,
-            1236,
-            1244,
-            1257,
-            1401,
-            1427,
-            1475,
-            1533,
-            1606,
-            1011
-        ]
-    },
-    {
-        "idx": 3,
-        "key": "housing_rebuilding_3",
-        "subcat_en": "Streamline rebuilding process",
-        "color": "#b98acb",
-        "comment_ids": [
-            367,
-            513,
-            614,
-            679,
-            870,
-            1013,
-            1156,
-            1309,
-            1328,
-            1355,
-            1433,
-            1443,
-            1481,
-            1502,
-            1523,
-            1531,
-            1546,
-            1592,
-            1639
-        ]
-    },
-    {
-        "idx": 4,
-        "key": "housing_rebuilding_4",
-        "subcat_en": "Temporary housing",
-        "color": "#ee8176",
-        "comment_ids": [
-            270,
-            467,
-            508,
-            556,
-            597,
-            662,
-            691,
-            730,
-            738,
-            813,
-            839,
-            860,
-            960,
-            969,
-            991,
-            1184,
-            1295,
-            1371,
-            1400,
-            1632,
-            1676,
-            1688
-        ]
-    },
-    {
-        "idx": 5,
-        "key": "housing_rebuilding_5",
-        "subcat_en": "Other",
-        "color": "lightgray",
-        "comment_ids": [
-            272,
-            289,
-            323,
-            412,
-            425,
-            446,
-            500,
-            535,
-            567,
-            588,
-            642,
-            649,
-            1084,
-            669,
-            711,
-            756,
-            871,
-            874,
-            903,
-            923,
-            924,
-            932,
-            934,
-            988,
-            1012,
-            1015,
-            1043,
-            1074,
-            1219,
-            1248,
-            1267,
-            1272,
-            1276,
-            1389,
-            1411,
-            1554,
-            1568,
-            261,
-            417,
-            275
-        ]
-    }
-]
