--- scripts/scatterplots/plot-legends/legend_infrastructure_restoration.json
@@ -1,184 +1,0 @@
-[
-    {
-        "idx": 1,
-        "key": "infrastructure_restoration_1",
-        "subcat_en": "Altadena infrastructure",
-        "color": "#5ed0b9",
-        "comment_ids": [
-            280,
-            577,
-            582,
-            590,
-            616,
-            989,
-            1008,
-            1009,
-            1010,
-            1096,
-            1282,
-            1310
-        ]
-    },
-    {
-        "idx": 2,
-        "key": "infrastructure_restoration_2",
-        "subcat_en": "Broadband access",
-        "color": "#70b6e5",
-        "comment_ids": [
-            380,
-            551,
-            656,
-            723,
-            961,
-            968,
-            1263,
-            1564
-        ]
-    },
-    {
-        "idx": 3,
-        "key": "infrastructure_restoration_3",
-        "subcat_en": "Climate resilient infrastructure ",
-        "color": "#b98acb",
-        "comment_ids": [
-            293,
-            322,
-            353,
-            448,
-            648,
-            787,
-            902,
-            1062,
-            1068,
-            1180,
-            1278,
-            1356,
-            1376,
-            1538,
-            1587,
-            1599,
-            1718
-        ]
-    },
-    {
-        "idx": 4,
-        "key": "infrastructure_restoration_4",
-        "subcat_en": "Underground power lines",
-        "color": "#ee8176",
-        "comment_ids": [
-            281,
-            359,
-            387,
-            399,
-            406,
-            422,
-            430,
-            433,
-            457,
-            507,
-            516,
-            517,
-            548,
-            563,
-            620,
-            628,
-            657,
-            696,
-            724,
-            732,
-            742,
-            784,
-            791,
-            802,
-            808,
-            830,
-            835,
-            838,
-            868,
-            882,
-            915,
-            929,
-            936,
-            973,
-            1006,
-            1039,
-            1042,
-            1058,
-            1087,
-            1150,
-            1159,
-            1166,
-            1175,
-            1187,
-            1192,
-            1197,
-            1213,
-            1221,
-            1224,
-            1225,
-            1242,
-            1250,
-            1258,
-            1259,
-            1280,
-            1322,
-            1329,
-            1331,
-            1343,
-            1361,
-            1391,
-            1393,
-            1402,
-            1407,
-            1413,
-            1416,
-            1423,
-            1428,
-            1451,
-            1461,
-            1466,
-            1476,
-            1480,
-            1483,
-            1503,
-            1504,
-            1515,
-            1536,
-            1541,
-            1559,
-            1574,
-            1618,
-            1624,
-            1626,
-            1637,
-            1640,
-            1643,
-            1657
-        ]
-    },
-    {
-        "idx": 5,
-        "key": "infrastructure_restoration_5",
-        "subcat_en": "Other",
-        "color": "lightgray",
-        "comment_ids": [
-            1005,
-            407,
-            596,
-            665,
-            686,
-            697,
-            767,
-            894,
-            956,
-            1149,
-            1265,
-            1299,
-            1431,
-            1530,
-            1593,
-            1605,
-            950
-        ]
-    }
-]
