--- src/public/data/legend_emergency_communication.json
@@ -1,146 +1,0 @@
-[
-    {
-        "idx": 1,
-        "key": "emergency_communication_1",
-        "subcat_en": "Emergency alert apps",
-        "color": "#5ed0b9",
-        "comment_ids": [
-            461,
-            619,
-            659,
-            750,
-            776,
-            810,
-            1026,
-            1040,
-            1081,
-            1094,
-            1176,
-            1346,
-            1390,
-            1425
-        ]
-    },
-    {
-        "idx": 2,
-        "key": "emergency_communication_2",
-        "subcat_en": "Emergency alert systems",
-        "color": "#70b6e5",
-        "comment_ids": [
-            361,
-            423,
-            498,
-            833,
-            879,
-            909,
-            931,
-            1157,
-            1324,
-            1336,
-            1363,
-            1366,
-            1378,
-            1638
-        ]
-    },
-    {
-        "idx": 3,
-        "key": "emergency_communication_3",
-        "subcat_en": "Post-emergency communications",
-        "color": "#b98acb",
-        "comment_ids": [
-            263,
-            266,
-            284,
-            291,
-            344,
-            358,
-            389,
-            426,
-            444,
-            525,
-            533,
-            538,
-            553,
-            566,
-            574,
-            595,
-            606,
-            650,
-            680,
-            729,
-            809,
-            837,
-            876,
-            877,
-            881,
-            959,
-            990,
-            1025,
-            1027,
-            1045,
-            1186,
-            1188,
-            1193,
-            1204,
-            1226,
-            1269,
-            1285,
-            1312,
-            1396,
-            1438,
-            1511,
-            1521,
-            1543,
-            1552,
-            1562,
-            1573,
-            1581,
-            1582,
-            1601,
-            1647,
-            1656,
-            1681,
-            1685,
-            1691,
-            1707,
-            1721
-        ]
-    },
-    {
-        "idx": 4,
-        "key": "emergency_communication_4",
-        "subcat_en": "Other",
-        "color": "lightgray",
-        "comment_ids": [
-            321,
-            427,
-            480,
-            518,
-            578,
-            636,
-            690,
-            744,
-            800,
-            804,
-            826,
-            880,
-            905,
-            941,
-            985,
-            1023,
-            1024,
-            1028,
-            1145,
-            1251,
-            1284,
-            1323,
-            1454,
-            1489,
-            1528,
-            1610,
-            1668,
-            1737
-        ]
-    }
-]
