--- src/public/data/legend_emergency_planning_safety.json
@@ -1,172 +1,0 @@
-[
-    {
-        "idx": 1,
-        "key": "emergency_planning_safety_1",
-        "subcat_en": "Emergency response improvements",
-        "color": "#5ed0b9",
-        "comment_ids": [
-            375,
-            481,
-            587,
-            601,
-            612,
-            664,
-            713,
-            761,
-            763,
-            807,
-            827,
-            859,
-            864,
-            1079,
-            1152,
-            1163,
-            1171,
-            1178,
-            1252,
-            1306,
-            1313,
-            1412,
-            1457,
-            1462,
-            1500,
-            1602,
-            1617,
-            1731
-        ]
-    },
-    {
-        "idx": 2,
-        "key": "emergency_planning_safety_2",
-        "subcat_en": "Emergency warning systems",
-        "color": "#70b6e5",
-        "comment_ids": [
-            555,
-            892,
-            935,
-            963,
-            1001,
-            1038,
-            1047,
-            1070,
-            1203,
-            1217,
-            1232,
-            1240,
-            1264,
-            1291,
-            1333,
-            1397,
-            1442,
-            1571,
-            1717
-        ]
-    },
-    {
-        "idx": 3,
-        "key": "emergency_planning_safety_3",
-        "subcat_en": "Evacuation planning",
-        "color": "#b98acb",
-        "comment_ids": [
-            308,
-            313,
-            328,
-            365,
-            398,
-            408,
-            418,
-            521,
-            666,
-            993,
-            1002,
-            1004,
-            1052,
-            1064,
-            1191,
-            1243,
-            1292,
-            1353,
-            1368,
-            1385,
-            1387,
-            1478,
-            1499,
-            1516,
-            1532,
-            1555,
-            1590,
-            1652,
-            1684
-        ]
-    },
-    {
-        "idx": 4,
-        "key": "emergency_planning_safety_4",
-        "subcat_en": "Next-gen wildfire readiness",
-        "color": "#ee8176",
-        "comment_ids": [
-            447,
-            449,
-            523,
-            701,
-            717,
-            878,
-            907,
-            911,
-            953,
-            980,
-            981,
-            1003,
-            1065,
-            1066,
-            1072,
-            1181,
-            1268,
-            1338,
-            1421,
-            1487,
-            1567,
-            1583,
-            1600,
-            1689,
-            1701,
-            1705
-        ]
-    },
-    {
-        "idx": 5,
-        "key": "emergency_planning_safety_5",
-        "subcat_en": "Other",
-        "color": "lightgray",
-        "comment_ids": [
-            295,
-            363,
-            376,
-            527,
-            554,
-            584,
-            605,
-            632,
-            633,
-            634,
-            635,
-            643,
-            681,
-            731,
-            782,
-            814,
-            979,
-            1085,
-            1153,
-            1235,
-            1274,
-            1297,
-            1304,
-            1434,
-            1449,
-            1622,
-            1667,
-            948
-        ]
-    }
-]
