--- src/public/data/legend_environmental_recovery.json
@@ -1,157 +1,0 @@
-[
-    {
-        "idx": 1,
-        "key": "environmental_recovery_1",
-        "subcat_en": "Altadena-specific concerns",
-        "color": "#5ed0b9",
-        "comment_ids": [
-            561,
-            460,
-            579,
-            1255,
-            1095
-        ]
-    },
-    {
-        "idx": 2,
-        "key": "environmental_recovery_2",
-        "subcat_en": "Habitat restoration and natural fire prevention",
-        "color": "#70b6e5",
-        "comment_ids": [
-            1739,
-            1609,
-            1055,
-            326,
-            382,
-            644,
-            674,
-            685,
-            688,
-            700,
-            708,
-            904,
-            1321,
-            1660,
-            1679,
-            1715
-        ]
-    },
-    {
-        "idx": 3,
-        "key": "environmental_recovery_3",
-        "subcat_en": "Satisfied with debris removal",
-        "color": "#b98acb",
-        "comment_ids": [
-            1682,
-            1671,
-            1540,
-            1410,
-            1522
-        ]
-    },
-    {
-        "idx": 4,
-        "key": "environmental_recovery_4",
-        "subcat_en": "Soil and other environmental testing",
-        "color": "#ee8176",
-        "comment_ids": [
-            715,
-            1360,
-            273,
-            285,
-            391,
-            600,
-            611,
-            629,
-            630,
-            779,
-            828,
-            856,
-            866,
-            1183,
-            1199,
-            1214,
-            1261,
-            1275,
-            1283,
-            1354,
-            1382,
-            1493,
-            1498,
-            1505,
-            1548,
-            1549,
-            1628,
-            1642,
-            1669,
-            1686,
-            1702
-        ]
-    },
-    {
-        "idx": 5,
-        "key": "environmental_recovery_5",
-        "subcat_en": "Other",
-        "color": "lightgray",
-        "comment_ids": [
-            1594,
-            1700,
-            294,
-            357,
-            599,
-            816,
-            1073,
-            1158,
-            1206,
-            1249,
-            1316,
-            1424,
-            1459,
-            1477,
-            1488,
-            1565,
-            1570,
-            1595,
-            1615,
-            434,
-            562,
-            683,
-            737,
-            772,
-            933,
-            992,
-            1296,
-            1388,
-            1711,
-            388,
-            429,
-            435,
-            524,
-            537,
-            560,
-            663,
-            675,
-            720,
-            721,
-            749,
-            771,
-            806,
-            872,
-            883,
-            938,
-            958,
-            972,
-            1046,
-            1053,
-            1071,
-            1148,
-            1155,
-            1256,
-            1330,
-            1367,
-            1439,
-            1452,
-            1557
-        ]
-    }
-]
