--- src/public/data/legend_financial_legal_assistance.json
@@ -1,166 +1,0 @@
-[
-    {
-        "idx": 1,
-        "key": "financial_legal_assistance_1",
-        "subcat_en": "Building code and permitting reform",
-        "color": "#5ed0b9",
-        "comment_ids": [
-            1699,
-            1692,
-            1633,
-            1607,
-            1473,
-            1444,
-            393,
-            547,
-            919,
-            1033,
-            951,
-            1031,
-            1029,
-            755,
-            735,
-            897,
-            509,
-            678,
-            781,
-            699,
-            832,
-            727,
-            1205,
-            920,
-            1173,
-            971,
-            1044,
-            545,
-            1245,
-            1341
-        ]
-    },
-    {
-        "idx": 2,
-        "key": "financial_legal_assistance_2",
-        "subcat_en": "Financial relief navigation",
-        "color": "#70b6e5",
-        "comment_ids": [
-            1708,
-            1723,
-            1683,
-            1645,
-            1616,
-            1509,
-            1627,
-            1436,
-            1351,
-            1399,
-            392,
-            978,
-            580,
-            257,
-            927,
-            706,
-            1358,
-            415,
-            823,
-            302,
-            733,
-            631,
-            1253,
-            1198,
-            1097,
-            1000,
-            1167,
-            1032,
-            899,
-            278,
-            940,
-            607,
-            269,
-            867,
-            526,
-            812,
-            707,
-            421,
-            564,
-            1211,
-            855,
-            747,
-            765,
-            888,
-            1246,
-            1326,
-            1305,
-            1288,
-            1289,
-            1339,
-            1576
-        ]
-    },
-    {
-        "idx": 3,
-        "key": "financial_legal_assistance_3",
-        "subcat_en": "Insurance reform",
-        "color": "#b98acb",
-        "comment_ids": [
-            1670,
-            1655,
-            1519,
-            1551,
-            1381,
-            1426,
-            1270,
-            384,
-            1458,
-            996,
-            805,
-            419,
-            1034,
-            462,
-            615,
-            519,
-            682,
-            777,
-            687,
-            591,
-            736,
-            949,
-            314,
-            1076,
-            1241,
-            1446,
-            1525
-        ]
-    },
-    {
-        "idx": 4,
-        "key": "financial_legal_assistance_4",
-        "subcat_en": "Other",
-        "color": "lightgray",
-        "comment_ids": [
-            1485,
-            1494,
-            1373,
-            610,
-            822,
-            1228,
-            885,
-            539,
-            440,
-            279,
-            298,
-            716,
-            824,
-            986,
-            1196,
-            1164,
-            1216,
-            1030,
-            840,
-            1151,
-            420,
-            1190,
-            1325,
-            1580
-        ]
-    }
-]
