--- src/public/data/legend_wildfire_prevention.json
@@ -1,160 +1,0 @@
-[
-    {
-        "idx": 1,
-        "key": "wildfire_prevention_1",
-        "subcat_en": "Accountable utilities and buried lines",
-        "color": "#5ed0b9",
-        "comment_ids": [
-            1169,
-            1732,
-            360,
-            510,
-            705,
-            764,
-            1041,
-            1060,
-            1090,
-            1208,
-            1238,
-            1277,
-            1311,
-            1335,
-            1345,
-            1365,
-            1409,
-            1448,
-            1463,
-            1534,
-            1673
-        ]
-    },
-    {
-        "idx": 2,
-        "key": "wildfire_prevention_2",
-        "subcat_en": "Controlled burn management",
-        "color": "#70b6e5",
-        "comment_ids": [
-            613,
-            818,
-            831,
-            900,
-            943,
-            1229,
-            1405,
-            1631,
-            1644,
-            886,
-            1709,
-            277
-        ]
-    },
-    {
-        "idx": 3,
-        "key": "wildfire_prevention_3",
-        "subcat_en": "Fire-safe home enforcement",
-        "color": "#b98acb",
-        "comment_ids": [
-            568,
-            655,
-            677,
-            889,
-            1020,
-            1144,
-            1370,
-            1675,
-            260,
-            623,
-            857,
-            895,
-            982,
-            1162,
-            1404,
-            1472,
-            1491,
-            1591
-        ]
-    },
-    {
-        "idx": 4,
-        "key": "wildfire_prevention_4",
-        "subcat_en": "Government fire prevention accountability",
-        "color": "#ee8176",
-        "comment_ids": [
-            1716,
-            983,
-            1298,
-            1520,
-            1566,
-            1604,
-            530,
-            754,
-            1501,
-            310,
-            325,
-            383,
-            397,
-            416,
-            456,
-            637,
-            710,
-            759,
-            785,
-            1017,
-            1037,
-            1379,
-            1398,
-            1484,
-            1577,
-            1620,
-            608,
-            757,
-            912,
-            913,
-            947,
-            964,
-            1016,
-            1018,
-            1054,
-            1082,
-            1083,
-            1281,
-            1344,
-            1432,
-            1529,
-            1561,
-            1585,
-            1693,
-            1696,
-            718,
-            364,
-            514,
-            558,
-            413,
-            528,
-            292,
-            1651
-        ]
-    },
-    {
-        "idx": 5,
-        "key": "wildfire_prevention_5",
-        "subcat_en": "Other",
-        "color": "lightgray",
-        "comment_ids": [
-            309,
-            366,
-            405,
-            494,
-            726,
-            811,
-            939,
-            1019,
-            1056,
-            1189,
-            1294,
-            1507,
-            1550,
-            1654
-        ]
-    }
-]
